requirements.txt
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
google-genai==1.56.0
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-google-genai==4.1.2
langchain-google-vertexai==3.2.0
# langchain-mongodb==0.9.0
langchain-ollama==1.0.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
# langgraph==1.0.5
# langgraph-checkpoint==3.0.1
# langgraph-checkpoint-mongodb==0.3.0
numpy==2.4.0
ollama==0.6.1
openai==2.11
pydantic==2.12.5
# pymongo==4.15.5
pypdf==6.4.2
qdrant-client==1.16.2
redis==7.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rq==2.6.1
rsa==4.9.1
sentry-sdk==2.47.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
speechrecognition==3.14.4
sqlalchemy==2.0.45
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
tenacity==9.1.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.2
uuid-utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1; platform_system != "Windows"
validators==0.35.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
rq==2.6.1
python-multipart==0.0.22
PyJWT==2.8.0
passlib==1.7.4
psycopg2-binary==2.9.11
bcrypt==5.0.0